add action.js
