+++ src/store/action.js
@@ -1,0 +1,5 @@
+export const ActionType = {
+  CHANGE_CITY: `main/changeCity`, // почему тут мейн?
+  FILTERED_OFFERS: `main/filteredOffers`
+};
+
